requirements.txt
# ══════════════════════════════════════════════════════════════
# Polymarket Bot v3.0 - Dependencies
# Author: Carlos David Donoso Cordero (ddchack)
# ══════════════════════════════════════════════════════════════
py-clob-client>=0.0.1
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
textblob>=0.17.1
requests>=2.31.0
websockets>=12.0
apscheduler>=3.10.0
scikit-learn>=1.3.0
web3>=6.0.0
